update (vesting-routes): multiple fixes, removed unused comp, cleanup. Rdy for review
