+++ src/popup/hive/pages/app-container/home/home.component.tsx
@@ -1,7 +1,6 @@
 import { UserLastCurrentRoutes } from '@interfaces/vesting-routes.interface';
 import { setSuccessMessage } from '@popup/hive/actions/message.actions';
 import { VestingRoutesPopupComponent } from '@popup/hive/pages/app-container/vesting-routes-popup/vesting-routes-popup.component';
-import { HiveTxUtils } from '@popup/hive/utils/hive-tx.utils';
 import { VestingRoutesUtils } from '@popup/hive/utils/vesting-routes.utils';
 import { LocalStorageKeyEnum } from '@reference-data/local-storage-key.enum';
 import { Screen } from '@reference-data/screen.enum';
@@ -68,86 +67,15 @@ const Home = ({
     initSurvey();
     initCheckKeysOnAccounts(accounts);
     initCheckVestingRoutes(accounts);
-    //TODO remove testing block bellow
-    // checkResultTxId('87c896e803062c1e13be777eabac01e29fcdbcb3');
-    // sendTestVestingRoutes(['keychain.tests', 'sexosentido'], 0);
+    //TODO Remove after testings & code review, this block bellow
+    // VestingRoutesUtils.sendTestVestingRoutes(
+    //   ['keychain.tests', 'sexosentido'],
+    //   accounts,
+    //   500,
+    //   true,
+    // );
     //end block
   }, []);
-
-  //TODo remove block bellow
-  const checkResultTxId = async (txId: string) => {
-    const resultTxID = await HiveTxUtils.getTransaction(txId);
-    console.log('Results of last Txid', { resultTxID });
-  };
-  const sendTestVestingRoutes = async (
-    testingAccounts: string[],
-    percent: number,
-  ) => {
-    testingAccounts.map(async (acc) => {
-      const userAK = accounts.find((a) => a.name === acc)?.keys.active;
-      if (userAK) {
-        const result = await VestingRoutesUtils.sendVestingRoute(
-          acc,
-          'theghost1980',
-          percent,
-          false,
-          userAK,
-        );
-        console.log('sendVestingRoute test results: ', { result });
-        const result2 = await VestingRoutesUtils.sendVestingRoute(
-          acc,
-          'stoodkev',
-          percent,
-          false,
-          userAK,
-        );
-        console.log('sendVestingRoute test results: ', { result2 });
-        const result3 = await VestingRoutesUtils.sendVestingRoute(
-          acc,
-          'sai.baba',
-          percent,
-          false,
-          userAK,
-        );
-        console.log('sendVestingRoute test results: ', { result3 });
-      } else {
-        Logger.error(`Need to add active key for: ${acc}`);
-      }
-    });
-    //TODO cleanup bellow
-    // const testingUsername = 'keychain.tests';
-    // const userAK = accounts.find((a) => a.name === testingUsername)?.keys
-    //   .active;
-    // if (userAK) {
-    //   const result = await VestingRoutesUtils.sendVestingRoute(
-    //     'keychain.tests',
-    //     'theghost1980',
-    //     percent,
-    //     false,
-    //     userAK,
-    //   );
-    //   console.log('sendVestingRoute test results: ', { result });
-    //   const result2 = await VestingRoutesUtils.sendVestingRoute(
-    //     'keychain.tests',
-    //     'stoodkev',
-    //     percent,
-    //     false,
-    //     userAK,
-    //   );
-    //   console.log('sendVestingRoute test results: ', { result2 });
-    //   const result3 = await VestingRoutesUtils.sendVestingRoute(
-    //     'keychain.tests',
-    //     'sexosentido',
-    //     percent,
-    //     false,
-    //     userAK,
-    //   );
-    //   console.log('sendVestingRoute test results: ', { result3 });
-    // } else {
-    //   Logger.error(`Need to add active key for: ${testingUsername}`);
-    // }
-  };
-  //end block
 
   useEffect(() => {
     if (activeRpc && activeRpc.uri !== 'NULL')
@@ -258,13 +186,11 @@ const Home = ({
       VestingRoutesUtils.saveLastVestingRoutes(currentVestingRoutes);
       return;
     } else {
-      console.log({ currentVestingRoutes, lastVestingRoutes }); //TODO remove line
       const wrongVestingRoutes = VestingRoutesUtils.getWrongVestingRoutes(
         lastVestingRoutes,
         currentVestingRoutes,
       );
       if (wrongVestingRoutes) {
-        console.log({ wrongVestingRoutes }); //TODO remove line
         setDisplayWrongVestingRoutesPopup(wrongVestingRoutes);
       }
     }
@@ -304,11 +230,6 @@ const Home = ({
           clearDisplayWrongVestingRoutes={() =>
             setDisplayWrongVestingRoutesPopup(undefined)
           }
-          //TODO bellow check & cleanup
-          // setDisplayWrongVestingRoutesPopup={(updated) =>
-          //   setDisplayWrongVestingRoutesPopup(updated)
-          // }
-          // localAccounts={accounts}
         />
       );
     }
